requirements.txt
py-clob-client>=0.34.0
anthropic>=0.39.0
httpx>=0.27.0
python-dotenv>=1.0.0
supabase>=2.0.0
aiohttp>=3.9.0
web3>=6.0.0
python-telegram-bot>=21.0
